requirements.txt
alembic==1.14.0
babel==2.16.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
elastic-transport==8.15.1
elasticsearch==8.17.0
email_validator==2.2.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
Mako==1.3.6
marisa-trie==1.2.1
MarkupSafe==3.0.2
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
